+++ software/contrib/radio_scanner.py
@@ -1,80 +1,106 @@
-from europi import *
+
+try:
+    # Local development
+    from software.firmware.europi import ain, b1, b2, clamp, cvs, din, k1, k2, oled
+    from software.firmware.europi import MAX_OUTPUT_VOLTAGE, OLED_HEIGHT, OLED_WIDTH
+    from software.firmware.europi_script import EuroPiScript
+except ImportError:
+    # Device import path
+    from europi import *
+    from europi_script import EuroPiScript
+
 from time import sleep_ms, ticks_diff, ticks_ms
 
 HEADER_DURATION = 2000  # 2 seconds in ms
 
 
-def remap_knob():
-    global knob_mapping
-    knob_mapping += 1
-    if knob_mapping == 3:
-        knob_mapping = 0
+class RadioScanner(EuroPiScript):
+    def __init__(self):
+        super().__init__()
 
-def rotate_cvs():
-    global cv_mapping
-    new_cv_mapping = []
-    for cv in cv_mapping[1:]: #Append all but the first (in order)
-        new_cv_mapping.append(cv)
-    new_cv_mapping.append(cv_mapping[0]) #Append the first
+        # Load state if previous state exists.
+        state = self.load_state_json()
+        # Set state variables with default fallback values if not found in the
+        # json save state.
+        self.knob_mapping = state.get("knob_mapping", 0)
+        self.cv_mapping = state.get("cv_mapping", [0, 1, 2, 3, 4, 5])
 
-    cv_mapping = new_cv_mapping
+        self.knob_mapping_text = ['Off', 'Knob 1', 'Knob 2']
 
-def value_to_cv(value):
-    return value * MAX_OUTPUT_VOLTAGE
+        def remap_knob():
+            self.knob_mapping += 1
+            if self.knob_mapping == 3:
+                self.knob_mapping = 0
+            self.save_state()
 
-def x_to_oled(x):
-    return round(x * (OLED_WIDTH - 1))
+        def rotate_cvs():
+            self.cv_mapping = self.cv_mapping[1:] + [self.cv_mapping[0]]
+            self.save_state()
 
-def y_to_oled(y):
-    return (OLED_HEIGHT - 1) - round(y * (OLED_HEIGHT - 1))
+        b1.handler(rotate_cvs)
+        din.handler(rotate_cvs)
+        b2.handler(remap_knob)
 
-def do_step(x, y):
-    oledx = x_to_oled(x)
-    oledy = y_to_oled(y)
+    def save_state(self):
+        """Save the current state variables as JSON."""
+        state = {
+            "knob_mapping": self.knob_mapping,
+            "cv_mapping": self.cv_mapping,
+        }
+        self.save_state_json(state)
 
-    cvx = value_to_cv(x)
-    cvy = value_to_cv(y)
+    def value_to_cv(self, value):
+        return value * MAX_OUTPUT_VOLTAGE
 
-    oled.fill(0)
-    oled.vline(oledx, 0, OLED_HEIGHT, 1)
-    oled.hline(0, oledy, OLED_WIDTH, 1)
+    def x_to_oled(self, x):
+        return round(x * (OLED_WIDTH - 1))
 
-    cv_mapping[0].voltage(cvx)  # 0 to 10 (volts)
-    cv_mapping[1].voltage(cvy)
-    cv_mapping[2].voltage(abs(cvy - cvx))
-    cv_mapping[3].voltage(MAX_OUTPUT_VOLTAGE - cvx)
-    cv_mapping[4].voltage(MAX_OUTPUT_VOLTAGE - cvy)
-    cv_mapping[5].voltage(MAX_OUTPUT_VOLTAGE - abs(cvy - cvx))
+    def y_to_oled(self, y):
+        return (OLED_HEIGHT - 1) - round(y * (OLED_HEIGHT - 1))
 
-    sleep_ms(10)
+    def do_step(self, x, y):
+        oledx = self.x_to_oled(x)
+        oledy = self.y_to_oled(y)
 
-def display_mapping(new_map):
-    oled.fill_rect(0, 0, 64, 12, 1)
-    oled.text(knob_mapping_text[new_map], 0, 2, 0)
+        cvx = self.value_to_cv(x)
+        cvy = self.value_to_cv(y)
 
+        oled.fill(0)
+        oled.vline(oledx, 0, OLED_HEIGHT, 1)
+        oled.hline(0, oledy, OLED_WIDTH, 1)
 
-cv_mapping = [cv1, cv2, cv3, cv4, cv5, cv6]
-knob_mapping = 0 #0 = not used, 1 = knob 1, 2 = knob 2
-knob_mapping_text = ['Off', 'Knob 1', 'Knob 2']
+        cvs[self.cv_mapping[0]].voltage(cvx)  # 0 to 10 (volts)
+        cvs[self.cv_mapping[1]].voltage(cvy)
+        cvs[self.cv_mapping[2]].voltage(abs(cvy - cvx))
+        cvs[self.cv_mapping[3]].voltage(MAX_OUTPUT_VOLTAGE - cvx)
+        cvs[self.cv_mapping[4]].voltage(MAX_OUTPUT_VOLTAGE - cvy)
+        cvs[self.cv_mapping[5]].voltage(MAX_OUTPUT_VOLTAGE - abs(cvy - cvx))
 
-b1.handler(rotate_cvs)
-din.handler(rotate_cvs)
-b2.handler(remap_knob)
+        sleep_ms(10)
 
-while True:
+    def display_mapping(self, new_map):
+        oled.fill_rect(0, 0, 64, 12, 1)
+        oled.text(self.knob_mapping_text[new_map], 0, 2, 0)
 
-    if knob_mapping != 1:
-        x = k1.percent()
-    else:
-        x = clamp(ain.percent() + k1.percent(), 0, 1)
+    def main(self):
 
-    if knob_mapping != 2:
-        y = k2.percent()
-    else:
-        y = clamp(ain.percent() + k2.percent(), 0, 1)
+        while True:
+            if self.knob_mapping != 1:
+                x = k1.percent()
+            else:
+                x = clamp(ain.percent() + k1.percent(), 0, 1)
 
-    do_step(x, y)
+            if self.knob_mapping != 2:
+                y = k2.percent()
+            else:
+                y = clamp(ain.percent() + k2.percent(), 0, 1)
 
-    if ticks_diff(ticks_ms(), b2.last_pressed()) < HEADER_DURATION:
-        display_mapping(knob_mapping)
-    oled.show()
+            self.do_step(x, y)
+
+            if ticks_diff(ticks_ms(), b2.last_pressed()) < HEADER_DURATION:
+                self.display_mapping(self.knob_mapping)
+
+            oled.show()
+
+if __name__ == "__main__":
+    RadioScanner().main()